fix: Properly draw vector shapes' arrowhead and set `VectorShape`'s startPoint and endPoint properties to public.
The property `theta` is only set once in the constructor, so when a
`VectorField` instance tries to rerender its vector shapes, `theta`
remains unchanged, thus the arrowheads are rendered like a mess. Moving
the logic of getting the `theta` to a method will dynamically change it
based on the vector shape's `endPoint`.